match: pl_800407C8 (unsigned int->float conversion)
xCEC/xCF0 are int fields converted to float for pl_CalculateAverage,
but the target uses the unsigned int->double magic path (2^52, no xoris).
Cast the field reads to (u32) at the conversion points without changing
the statement structure, preserving the matched register/stack layout.

+++ src/melee/pl/plbonuslib.c
@@ -1290,8 +1290,8 @@ float pl_800407C8(int arg0)
     temp_r3 = Player_GetStaleMoveTableIndexPtr2(arg0);
     temp_r5 = temp_r3->x0_staleMoveTable.xCEC;
     if (temp_r5 != 0) {
-        float tmp = temp_r3->x0_staleMoveTable.xCEC;
-        temp_f31 = temp_r3->x0_staleMoveTable.xCF0;
+        float tmp = (u32) temp_r3->x0_staleMoveTable.xCEC;
+        temp_f31 = (u32) temp_r3->x0_staleMoveTable.xCF0;
         return pl_CalculateAverage(temp_f31, tmp);
     }
     return 0.0f;
